+++ apps/gizual-app/src/primitives/canvas/canvas.context.ts
@@ -18,7 +18,7 @@ type CanvasContextProps = {
   debugLayout: boolean;
 };
 
-const CanvasContext = React.createContext({ useBlocks: useBlocks, debugLayout: false });
+const CanvasContext = React.createContext({} as CanvasContextProps);
 
 export { CanvasContext };
 export type { CanvasContextProps };